+++ R/bm_bitmap.R
@@ -1,15 +1,17 @@
-#' Bitmap object
+#' Bittermelon bitmap matrix object
 #'
-#' `bm_bitmap()` creates an S3 object representing bitmap.
+#' `bm_bitmap()` creates an S3 matrix subclass representing a bitmap.
 #'
-#' Bitmaps are represented as integer matrices with special class methods.
-#' The bottom left pixel is represented by the first row and first column.
-#' The bottom right pixel is represented by the first row and last column.
-#' The top left pixel is represented by the last row and first column.
-#' The top right pixel is represented by the last row and last column.
-#' Color bitmaps are supported (the integer can be any non-negative integer)
-#' but we are unlikely to ever support exporting color bitmap fonts.
-#' Color bitmaps can be cast to black-and-white bitmaps via [bm_clamp()].
+#' * Intended to represent binary bitmaps especially (but not limited to) bitmap font glyphs.
+#' * Bitmaps are represented as integer matrices with special class methods.
+#' * The bottom left pixel is represented by the first row and first column.
+#' * The bottom right pixel is represented by the first row and last column.
+#' * The top left pixel is represented by the last row and first column.
+#' * The top right pixel is represented by the last row and last column.
+#' * Non-binary bitmaps are supported (the integer can be any non-negative integer)
+#'   but we are unlikely to ever support exporting color bitmap fonts.
+#' * Non-binary bitmaps can be cast to binary bitmaps via [bm_clamp()].
+#' * See [bm_pixmap()] for an alternative S3 object backed by a color string matrix.
 #'
 #' @section Supported S3 methods:
 #'
@@ -24,7 +26,7 @@
 #' @param x Object to be converted to `bm_bitmap()`.
 #'   If not already an integer matrix it will be cast to one
 #'   by [as_bm_bitmap()].
-#' @return An integer matrix with a \dQuote{bm_bitmap} subclass.
+#' @return An integer matrix with \dQuote{bm_bitmap} and \dQuote{bm_matrix} subclasses.
 #' @examples
 #'  space <- bm_bitmap(matrix(0, nrow = 16, ncol = 16))
 #'  print(space)
@@ -36,18 +38,3 @@ bm_bitmap <- function(x) {
     else
         as_bm_bitmap(x)
 }
-
-#' Test if the object is a bitmap glyph object
-#'
-#' `is_bm_bitmap()` returns `TRUE` for `bm_bitmap` objects (or subclasses)
-#' and `FALSE` for all other objects.
-#' @param x An object
-#' @return `TRUE` or `FALSE`
-#' @examples
-#'  space_matrix <- matrix(0L, nrow = 16L, ncol = 16L)
-#'  is_bm_bitmap(space_matrix)
-#'  space_glyph <- bm_bitmap(space_matrix)
-#'  is_bm_bitmap(space_glyph)
-#' @seealso [bm_bitmap()]
-#' @export
-is_bm_bitmap <- function(x) inherits(x, "bm_bitmap")