+++ R/bm_font.R
@@ -22,7 +22,7 @@
 #'  R_glyph <- font[[str2ucp("R")]]
 #'  print(R_glyph)
 #' @return A named list with a \dQuote{bm_font} subclass.
-#' @seealso [is_bm_font()], [as_bm_font(), [hex2ucp()]]
+#' @seealso [is_bm_font()], [as_bm_font()], [hex2ucp()]
 #' @export
 bm_font <- function(x = bm_list(), comments = NULL, properties = NULL) {
     if (is_bm_font(x))